apps/zero/backend/flask/example/requirements.txt
requests
flask
pyngrok
flask_socketio
gunicorn
gevent
flask_sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
translatepy
flask_cors